add curl to routes to be have cleaner code
